requirement.txt
certifi==2018.4.16
chardet==3.0.4
decorator==4.3.0
Django==2.0.6
django-filter==1.1.0
djangorestframework==3.8.2
ez-setup==0.9
idna==2.7
imageio==2.3.0
Markdown==2.6.11
moviepy==0.2.3.5
numpy==1.14.5
Pillow==5.1.0
pytz==2018.4
requests==2.19.1
tqdm==4.23.4
urllib3==1.23
